add:fix: build timings, icon4py lib

+++ build_utils/full_build.sh
@@ -9,6 +9,12 @@
 
 set -e
 
+elpased(){
+    local seconds=$((stop - start))
+    printf '%02d:%02d:%02d\n' $((seconds/3600)) $((seconds%3600/60)) $((seconds%60))
+}
+
+
 BUILD_TYPE="${BUILD_TYPE:-SPACK}"
 GPU_MODE="${GPU_MODE:-py-substitute}"
 CAO_BUILD_DIR="${CAO_BUILD_DIR:-/dev/shm/${USER}/coupled_setup}"
@@ -44,8 +50,10 @@ fi
 # Get script dir
 # --------------
 if [ -n "${SLURM_JOB_ID:-}" ]; then
+    ON_COMPUTE_NODE="true"
     SCRIPT_PATH=$(scontrol show job "${SLURM_JOB_ID}" | awk -F= '/Command=/{print $2}')
 else
+    ON_COMPUTE_NODE="false"
     SCRIPT_PATH=$(realpath "${BASH_SOURCE[0]}")
 fi
 SCRIPT_PATH="$(realpath "${SCRIPT_PATH}")"
@@ -58,7 +66,9 @@ pushd "${CAO_BUILD_DIR}" 2>&1 >/dev/null
 
 # Get ICON
 # --------
+start=$(date +%s)
 echo "[CAO build] ... Getting ICON"
+
 CAO_ICON_REPO='git@github.com:C2SM/icon-exclaim.git'
 CAO_ICON_BRANCH='icon-dsl'
 CAO_ICON_COMMIT='5c5b742a969af2bd491e26cd0a05a35838f121c4'
@@ -83,10 +93,16 @@ else
 fi
 
 echo "[CAO build] ...... Getting icon4py"
-git clone -b "${CAO_ICON4PY_BRANCH}" "${CAO_ICON4PY_REPO}" "${CAO_ICON4PY_DIR}"
-pushd "${CAO_ICON4PY_DIR}" >/dev/null 2>&1
-git reset --hard "${CAO_ICON4PY_COMMIT}"
-popd >/dev/null 2>&1
+if [ -n "${CAO_ICON4PY_COMMIT}" ]; then
+    git clone -b "${CAO_ICON4PY_BRANCH}" "${CAO_ICON4PY_REPO}" "${CAO_ICON4PY_DIR}"
+    pushd "${CAO_ICON4PY_DIR}" >/dev/null 2>&1
+    git reset --hard "${CAO_ICON4PY_COMMIT}"
+    popd >/dev/null 2>&1
+else
+    git clone --depth 1 -b "${CAO_ICON4PY_BRANCH}" "${CAO_ICON4PY_REPO}" "${CAO_ICON4PY_DIR}"
+fi
+stop=$(date +%s)
+echo "[CAO build] ... Getting ICON => done in $(elpased)"
 
 # Apply patches
 # -------------
@@ -100,7 +116,9 @@ popd >/dev/null 2>&1
 
 # Build
 # -----
+start=$(date +%s)
 echo "[CAO build] ... Building ICON"
+
 CPU_BUILD_DIR="${CAO_ICON_DIR}/build-cpu"
 GPU_BUILD_DIR="${CAO_ICON_DIR}/build-gpu-${GPU_MODE}" 
 mkdir -p ${CPU_BUILD_DIR} ${GPU_BUILD_DIR}
@@ -113,7 +131,7 @@ if [ "${BUILD_TYPE}" ==  "SPACK" ]; then
     if [ "${build_cpu}" == "true" ]; then
         echo "[CAO build] ...... Building cpu"
         pushd "${CPU_BUILD_DIR}" >/dev/null 2>&1
-        uenv run ${UENV} --view default -- ../config/cscs/santis.cpu.nvhpc
+        uenv run ${UENV} --view default -- time ../config/cscs/santis.cpu.nvhpc
         popd >/dev/null 2>&1
     fi
 
@@ -121,9 +139,9 @@ if [ "${BUILD_TYPE}" ==  "SPACK" ]; then
         echo "[CAO build] ...... Building gpu-${GPU_MODE}"
         pushd "${GPU_BUILD_DIR}" >/dev/null 2>&1
         if [ "${GPU_MODE}" == "acc" ]; then
-            uenv run ${UENV} --view default -- ../config/cscs/santis.gpu.nvhpc
+            uenv run ${UENV} --view default -- time ../config/cscs/santis.gpu.nvhpc
         elif [ "${GPU_MODE}" == "py-substitute" ]; then
-            uenv run ${UENV} --view default -- ../config/cscs/santis.gpu.nvhpc.py.substitute
+            uenv run ${UENV} --view default -- time ../config/cscs/santis.gpu.nvhpc.py.substitute
         else
             echo "[CAO build] ERROR: unknown GPU_MODE ${GPU_MODE}"
             exit 1
@@ -158,14 +176,25 @@ else
     
 fi
 
-popd >/dev/null 2>&1
+stop=$(date +%s)
+echo "[CAO build] ... Building ICON => done in $(elpased)"
 
 # Retreive and clean
 # ------------------
+start=$(date +%s)
 echo "[CAO build] ... retreiving build from ${CAO_BUILD_DIR}"
-rsync -a --delete "${CAO_BUILD_DIR}/${CAO_ICON_DIR}/" "${CAO_ICON_DIR}/"
+
+popd >/dev/null 2>&1
+time rsync -a --delete "${CAO_BUILD_DIR}/${CAO_ICON_DIR}/" "${CAO_ICON_DIR}/"
+
+stop=$(date +%s)
+echo "[CAO build] ... retreiving => done in $(elpased)"
 
 echo "[CAO build] ... cleaning ${CAO_BUILD_DIR}"
 rm -rf "${CAO_BUILD_DIR}/${CAO_ICON_DIR}"
 
 echo "[CAO build] ... build complete"
+
+if [ "${ON_COMPUTE_NODE}" == "true" ]; then
+    sacct -j "${SLURM_JOB_ID}" --format "JobID, JobName, AllocCPUs, Elapsed, ElapsedRaw, CPUTimeRAW, ConsumedEnergyRaw, MaxRSS, MaxVMSize, AveRSS"
+fi
